Data Integrity and Authentication/multi-factor-authentication-2fa-system/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.1
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
Pillow>=10.3.0
mysql-connector-python==8.2.0
